requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image processing and computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
imageio>=2.31.0

# Machine learning and data science
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep learning utilities
albumentations>=1.3.0
timm>=0.9.0
torchmetrics>=1.0.0

# Evaluation metrics
lpips>=0.1.4
pytorch-msssim>=1.0.0

# Model optimization and export
onnx>=1.14.0
onnxruntime>=1.15.0
coremltools>=7.0.0

# Dashboard and visualization
streamlit>=1.25.0
gradio>=3.40.0
plotly>=5.15.0
bokeh>=3.2.0

# Data handling
pandas>=2.0.0
h5py>=3.9.0
tqdm>=4.65.0

# Configuration and utilities
pyyaml>=6.0
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.7.0
flake8>=6.0.0
pytest>=7.4.0

# Optional GPU acceleration (uncomment if using CUDA)
# torch==2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.15.0+cu118 --index-url https://download.pytorch.org/whl/cu118
